DEADSGOLD/client_app/requirements.txt
absl-py==1.4.0
accelerate==0.24.1
aio-pika==8.2.3
aiobotocore==2.8.0
aiofiles==23.2.1
aiogram==2.25.2
aiohttp==3.8.6
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiormq==6.4.2
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.9.1.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.0.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
backports.zoneinfo==0.2.1
Batman==0.7.0
beautifulsoup4==4.12.2
bidict==0.22.1
bitarray==2.9.0
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
blosc2==2.0.0
bnm==1.0.0
boto3==1.29.2
botocore==1.32.7
Bottleneck==1.3.7
brotlipy==0.7.0
BTM==1.1.5
CacheControl==0.12.14
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
cloudpathlib==0.16.0
cloudpickle==2.2.1
cmake==3.28.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
comm==0.2.0
confection==0.1.3
confluent-kafka==2.3.0
contourpy==1.1.1
cramjam==2.7.0
cryptography==41.0.5
cycler==0.12.1
cymem==2.0.8
Cython==3.0.6
cytoolz==0.12.2
dask==2022.10.2
dataclasses-json==0.6.2
db-dtypes==1.1.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
discord==2.3.2
discord.py==2.3.2
diskcache==5.6.3
distlib==0.3.8
distro==1.8.0
dnspython==2.3.0
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.7.0
eth-keys==0.4.0
eth-rlp==1.0.0
eth-typing==3.5.2
eth-utils==2.3.1
exceptiongroup==1.1.3
execnet==2.0.2
executing==2.0.1
fairscale==0.4.13
fastjsonschema==2.19.0
fastparquet==2023.10.1
fbmessenger==6.0.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.44.3
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.12.0
future==0.18.3
gast==0.4.0
gcsfs==2023.12.0
gitdb==4.0.11
GitPython==3.1.40
google==3.0.0
google-api-core==2.14.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.13.0
google-cloud-bigquery-storage==2.23.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
grpcio-status==1.59.3
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.10.0
hexbytes==0.3.1
html5lib==1.1
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.19.4
humanfriendly==10.0
hypothesis==6.91.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipykernel==6.27.1
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jax==0.4.13
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.12.0
kiwisolver==1.4.5
langchain==0.0.330
langcodes==3.3.0
langsmith==0.0.57
learn==1.0.0
libclang==16.0.6
-e git+https://github.com/facebookresearch/llama@ef351e9cd9496c579bf9f2bb036ef11bdc5ca3d2#egg=llama
llama_cpp_python==0.2.32
llm==0.12
llm-llama-cpp==0.3
llvmlite==0.41.1
locket==1.0.0
lru-dict==1.2.0
lxml==4.9.3
magic-filter==1.0.12
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mattermostwrapper==2.2
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==5.2.0
murmurhash==1.0.10
mypy-extensions==1.0.0
natsort==8.4.0
nbclient==0.7.4
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==2.6.3
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numexpr==2.8.6
numpy==1.23.5
oauthlib==3.2.2
odfpy==1.4.1
openai==1.9.0
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pamqp==3.2.1
pandas==2.0.3
pandas-gbq==0.19.2
pandocfilters==1.5.0
parsimonious==0.9.0
parso==0.8.3
partd==1.4.1
pbr==6.0.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
pluggy==1.3.0
ply==3.10
portalocker==2.8.2
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.22.3
protobuf==4.23.3
psutil==5.9.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.9
pydantic_core==2.10.1
pydata-google-auth==1.8.2
pydeck==0.8.1b0
pydot==1.4.2
pydub==0.25.1
pygame==2.5.2
Pygments==2.16.1
PyJWT==2.8.0
pykwalify==1.8.0
pylint==3.0.2
pymongo==4.3.3
PyMySQL==1.1.0
pyomega==0.1.17
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
pyreadstat==1.2.5
pyrsistent==0.20.0
pysigma==0.8.12
pytango==9.4.1
pytest==7.4.3
pytest-asyncio==0.23.0
pytest-xdist==3.5.0
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-snappy==0.6.1
python-socketio==5.10.0
python-ulid==1.1.0
pytz==2022.7.1
pyunormalize==15.1.0
pywin32==306
pywinpty==2.0.12
pyxlsb==1.0.10
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
questionary==1.10.0
randomname==0.1.5
rasa==3.6.13
rasa-sdk==3.6.2
redis==4.6.0
referencing==0.32.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rlp==4.0.0
rocketchat-API==1.30.0
rpds-py==0.13.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3fs==2023.12.0
s3transfer==0.7.0
safetensors==0.4.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.14.0
shellingham==1.5.4
sigma==0.0.1
simple-websocket==1.0.0
six==1.16.0
sk==0.0.1
sklearn==0.0
sklearn-crfsuite==0.3.6
slack-sdk==3.23.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-lookups-data==1.0.5
SQLAlchemy==1.4.50
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.36
srsly==2.4.8
stack-data==0.6.3
starlette==0.36.2
statesman==1.0.4
stevedore==5.1.0
streamlit==1.28.2
structlog==23.2.0
structlog-sentry==2.0.3
sudo==1.0.0
sympy==1.12
tables==3.8.0
tabulate==0.9.0
tarsafe==0.0.4
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.18.0
terminaltables==3.1.10
tf==1.0.0
Theano==1.0.5
thinc==8.2.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.0
transformers==4.36.2
twilio==8.2.2
typer==0.9.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing-utils==0.1.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.24.0.post1
validators==0.22.0
virtualenv==20.25.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.0
voila==0.5.5
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.9
weasel==0.3.4
web3==6.12.0
webcolors==1.13
webencodings==0.5.1
webexteamssdk==1.6.1
websocket-client==1.7.0
websockets==10.4
Werkzeug==3.0.1
wget==3.2
widgetsnbextension==4.0.9
wrapt==1.14.1
wsproto==1.2.0
xarray==2023.1.0
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.2
zipp==3.17.0
zstandard==0.22.0
solana==0.29.0
spl-token==0.2.0